# Optimize reading strings across segments