Before
```
|                                    Method | BytesToParse | encodedSize |       Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------------------------ |------------- |------------ |-----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
|              ParseString_CodedInputStream |        10080 |           1 |  64.871 us | 0.2090 us | 0.1745 us |      0.1221 |           - |           - |               176 B |
|                  ParseString_ParseContext |        10080 |           1 |  66.493 us | 0.8300 us | 0.6931 us |           - |           - |           - |                   - |
| ParseString_ParseContext_MultipleSegments |        10080 |           1 |  80.660 us | 1.5452 us | 1.7175 us |           - |           - |           - |                   - |
|              ParseString_CodedInputStream |        10080 |           4 |  80.335 us | 1.2552 us | 1.1127 us |    102.6611 |           - |           - |             80816 B |
|                  ParseString_ParseContext |        10080 |           4 | 106.574 us | 0.9521 us | 0.8906 us |    102.4170 |           - |           - |             80640 B |
| ParseString_ParseContext_MultipleSegments |        10080 |           4 | 110.729 us | 2.6566 us | 2.4850 us |    102.4170 |           - |           - |             80640 B |
|              ParseString_CodedInputStream |        10080 |          10 |  42.858 us | 0.6439 us | 0.6023 us |     61.7065 |           - |           - |             48560 B |
|                  ParseString_ParseContext |        10080 |          10 |  54.975 us | 0.1663 us | 0.1474 us |     61.4624 |           - |           - |             48384 B |
| ParseString_ParseContext_MultipleSegments |        10080 |          10 |  70.892 us | 0.3835 us | 0.3399 us |     76.7822 |           - |           - |             60480 B |
|              ParseString_CodedInputStream |        10080 |         105 |  10.244 us | 0.0523 us | 0.0490 us |     29.5105 |           - |           - |             23216 B |
|                  ParseString_ParseContext |        10080 |         105 |  10.740 us | 0.1339 us | 0.1187 us |     29.2816 |           - |           - |             23040 B |
| ParseString_ParseContext_MultipleSegments |        10080 |         105 |  29.147 us | 0.1317 us | 0.1100 us |     51.7273 |           - |           - |             40704 B |
|              ParseString_CodedInputStream |        10080 |       10080 |   6.136 us | 0.1177 us | 0.1309 us |     25.6348 |           - |           - |             20360 B |
|                  ParseString_ParseContext |        10080 |       10080 |   5.732 us | 0.0272 us | 0.0241 us |     25.6348 |           - |           - |             20184 B |
| ParseString_ParseContext_MultipleSegments |        10080 |       10080 |  15.753 us | 0.1460 us | 0.1295 us |     38.4521 |           - |           - |             30344 B |
```

After
```
|                                    Method | BytesToParse | encodedSize |       Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------------------------ |------------- |------------ |-----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
|              ParseString_CodedInputStream |        10080 |           1 |  65.621 us | 0.3992 us | 0.3539 us |      0.1221 |           - |           - |               184 B |
|                  ParseString_ParseContext |        10080 |           1 |  65.119 us | 1.1548 us | 1.0237 us |           - |           - |           - |                   - |
| ParseString_ParseContext_MultipleSegments |        10080 |           1 |  76.722 us | 0.6850 us | 0.6073 us |           - |           - |           - |                   - |
|              ParseString_CodedInputStream |        10080 |           4 |  84.192 us | 0.7777 us | 0.7274 us |    102.6611 |           - |           - |             80824 B |
|                  ParseString_ParseContext |        10080 |           4 | 106.247 us | 1.1588 us | 1.0840 us |    102.4170 |           - |           - |             80640 B |
| ParseString_ParseContext_MultipleSegments |        10080 |           4 | 109.838 us | 0.4112 us | 0.3645 us |    102.4170 |           - |           - |             80640 B |
|              ParseString_CodedInputStream |        10080 |          10 |  43.997 us | 0.6887 us | 0.6105 us |     61.7065 |           - |           - |             48568 B |
|                  ParseString_ParseContext |        10080 |          10 |  55.625 us | 0.3236 us | 0.2868 us |     61.4624 |           - |           - |             48384 B |
| ParseString_ParseContext_MultipleSegments |        10080 |          10 |  78.260 us | 1.0150 us | 0.8476 us |     61.5234 |           - |           - |             48440 B |
|              ParseString_CodedInputStream |        10080 |         105 |  10.687 us | 0.2087 us | 0.2786 us |     29.5258 |           - |           - |             23224 B |
|                  ParseString_ParseContext |        10080 |         105 |  10.966 us | 0.2136 us | 0.1894 us |     29.2816 |           - |           - |             23040 B |
| ParseString_ParseContext_MultipleSegments |        10080 |         105 |  35.011 us | 0.4347 us | 0.4066 us |     29.3579 |           - |           - |             23096 B |
|              ParseString_CodedInputStream |        10080 |       10080 |   6.284 us | 0.1229 us | 0.1463 us |     25.6348 |           - |           - |             20368 B |
|                  ParseString_ParseContext |        10080 |       10080 |   5.795 us | 0.0716 us | 0.0670 us |     25.6348 |           - |           - |             20184 B |
| ParseString_ParseContext_MultipleSegments |        10080 |       10080 |  19.453 us | 0.2476 us | 0.2316 us |     25.6348 |           - |           - |             20240 B |
```