@JamesNK 2020-06-30
Updated numbers:

```
|                                    Method | BytesToParse | encodedSize |      Mean |     Error |    StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------------------------ |------------- |------------ |----------:|----------:|----------:|------------:|------------:|------------:|--------------------:|
|              ParseString_CodedInputStream |        10080 |           1 | 65.246 us | 0.7903 us | 0.6599 us |      0.1221 |           - |           - |               176 B |
|                  ParseString_ParseContext |        10080 |           1 | 65.650 us | 1.4109 us | 1.9313 us |           - |           - |           - |                   - |
|              ParseString_CodedInputStream |        10080 |           4 | 83.324 us | 1.4563 us | 1.2161 us |    102.6611 |           - |           - |             80816 B |
|                  ParseString_ParseContext |        10080 |           4 | 81.609 us | 1.6136 us | 2.9506 us |    102.4170 |           - |           - |             80640 B |
|              ParseString_CodedInputStream |        10080 |          10 | 41.770 us | 0.2447 us | 0.2169 us |     61.7065 |           - |           - |             48560 B |
|                  ParseString_ParseContext |        10080 |          10 | 40.760 us | 0.7884 us | 0.6989 us |     61.4624 |           - |           - |             48384 B |
|              ParseString_CodedInputStream |        10080 |         105 | 10.801 us | 0.2133 us | 0.2773 us |     29.5105 |           - |           - |             23216 B |
|                  ParseString_ParseContext |        10080 |         105 | 10.385 us | 0.1335 us | 0.1249 us |     29.2816 |           - |           - |             23040 B |
| ParseString_ParseContext_MultipleSegments |        10080 |         105 | 19.579 us | 0.0907 us | 0.0849 us |     29.2664 |           - |           - |             23040 B |
|              ParseString_CodedInputStream |        10080 |       10080 |  6.538 us | 0.1290 us | 0.1380 us |     25.6348 |           - |           - |             20360 B |
|                  ParseString_ParseContext |        10080 |       10080 |  6.112 us | 0.0757 us | 0.0671 us |     25.6348 |           - |           - |             20184 B |
| ParseString_ParseContext_MultipleSegments |        10080 |       10080 |  9.524 us | 0.1890 us | 0.2457 us |     25.6348 |           - |           - |             20184 B |
```